use string as key and extract util function
